program_files/requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argcomplete==1.8.1
argon2-cffi==21.1.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contextlib2==21.6.0
cryptography==3.4.8
cupshelpers==1.0
cvxopt==1.3.2
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.9.0
fonttools==4.47.0
fpdf2==2.7.7
frozenlist==1.2.0
fs==2.4.12
fsspec==2023.4.0
gast==0.5.2
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
idna==3.3
importlib-metadata==4.6.4
internetarchive==3.6.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
livereload==2.6.3
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mkdocs==1.1.2
more-itertools==8.10.0
mpmath==0.0.0
multidict==5.1.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.11.0
networkx==3.0
notebook==6.4.8
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
openai==1.6.1
packaging==21.3
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.1
pdfly==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipx==1.0.0
ply==3.11
pop-transition==1.1.2
powerline-status==2.8.2
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
proton-core==0.1.16
proton-keyring-linux==0.0.1
proton-keyring-linux-secretservice==0.0.1
proton-vpn-api-core==0.20.3
proton-vpn-connection==0.11.3
proton-vpn-gtk-app==4.1.10
proton-vpn-killswitch==0.2.0
proton-vpn-killswitch-network-manager==0.2.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.3.3
proton-vpn-network-manager-openvpn==0.0.4
proton-vpn-session==0.6.5
psutil==5.9.0
ptyprocess==0.7.0
PuLP==2.8.0
py==1.10.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==3.17.4
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
rich==13.7.0
schema==0.7.5
scikit-learn==1.4.0
scipy==1.8.0
seaborn==0.13.1
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
ssh-import-id==5.11
sympy==1.9
systemd-python==234
tabulate==0.8.9
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.2.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
urllib3==1.26.5
userpath==1.8.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
